fix(ci): complete CI green on top of contributor's fixes
Layers the remaining fixes onto the contributor's "fix ci issues" commit,
which addressed events.rs and the TypeScript error codes but left the rest.

- migrations: fix corrupted CREATE EXTENSION (0001) and drop CONCURRENTLY
  (0009) since sqlx runs each migration in a transaction
- Docker: commit Cargo.lock (was gitignored) and bump both Rust builder
  images rust:1.78 -> rust:1; build the Go API root package (.) not ./...
- Go: fix all golangci-lint errcheck findings, replace deprecated
  grpc.DialContext with grpc.NewClient; pin action v6 + golangci v1.64.8
- Rust unit job: require_services! skips (not panics) when TEST_DATABASE_URL
  is absent; set DATABASE_URL and --test-threads=1 for sqlx::test + config
- Rust integration job: add DATABASE_URL for sqlx::test; fix streamer
  poll_once tests (retry backoff cap, ledger-lag recompute, getLedgers vs
  getEvents mock separation)
- status.go: call Stat().TotalConns() instead of casting the method value
- validation/stats_test.go: fix typed-nil assert.NoError on *ValidationError
- regenerate api-types.gen.ts so the OpenAPI drift check passes

+++ services/api/handlers/status.go
@@ -100,7 +100,7 @@ func InternalStatus() http.HandlerFunc {
 
 			// DB pool connections
 			if statusDeps.db != nil {
-				resp.DBPoolOpenConns = int32(statusDeps.db.Stat().TotalConns)
+				resp.DBPoolOpenConns = statusDeps.db.Stat().TotalConns()
 			}
 
 			// Parse errors in last 24 hours
